reorganizing parameters and datasets, and update tiled dataset to use tiled client
