+++ src/predict_model.py
@@ -12,7 +12,8 @@ import torch
 from PIL import Image
 
 from helper_utils import embed_imgs, get_dataloaders
-from model import Autoencoder, TestingParameters
+from model import Autoencoder
+from parameters import InferenceParameters
 
 num_cpus = multiprocessing.cpu_count()
 if num_cpus > 6:
@@ -35,46 +36,48 @@ if __name__ == "__main__":
     parser.add_argument("-o", "--output_dir", help="output directory")
     parser.add_argument("-p", "--parameters", help="list of training parameters")
     args = parser.parse_args()
-    test_parameters = TestingParameters(**json.loads(args.parameters))
+    inference_parameters = InferenceParameters(**json.loads(args.parameters))
 
     device = (
         torch.device("cuda:0") if torch.cuda.is_available() else torch.device("cpu")
     )
     print("Device:", device)
 
-    if test_parameters.target_width * test_parameters.target_height > 0:
-        target_size = (test_parameters.target_width, test_parameters.target_height)
+    if inference_parameters.target_width * inference_parameters.target_height > 0:
+        target_size = (
+            inference_parameters.target_width,
+            inference_parameters.target_height,
+        )
     else:
         target_size = None
 
-    test_loader, (temp_channels, temp_w, temp_h), datasets_uris = get_dataloaders(
+    inference_loader, (temp_channels, temp_w, temp_h) = get_dataloaders(
         args.data_info,
-        test_parameters.batch_size,
+        inference_parameters.batch_size,
         NUM_WORKERS,
         shuffle=False,
         target_size=target_size,
-        log=test_parameters.log,
+        log=inference_parameters.log,
         train=False,
     )
 
     model = Autoencoder.load_from_checkpoint(args.model_dir + "/last.ckpt")
 
-    # Get latent space representation of test images and reconstructed images
-    test_img_embeds, test_result = embed_imgs(model, test_loader)
+    # Get latent space representation of inference images and reconstructed images
+    inference_img_embeds, inference_result = embed_imgs(model, inference_loader)
 
     # Create output directory if it does not exist
     output_dir = Path(args.output_dir)
     output_dir.mkdir(parents=True, exist_ok=True)
 
-    # Save latent space representation of test images
-    df = pd.DataFrame(test_img_embeds.cpu().detach().numpy())
-    df.index = datasets_uris
+    # Save latent space representation of inference images
+    df = pd.DataFrame(inference_img_embeds.cpu().detach().numpy())
     df.columns = df.columns.astype(str)
     df.to_parquet(f"{args.output_dir}/f_vectors.parquet", engine="pyarrow")
 
     # Reconstructed images
-    test_result = einops.rearrange(test_result, "n c x y -> n x y c")
-    test_result = test_result.cpu().detach().numpy()
+    inference_result = einops.rearrange(inference_result, "n c x y -> n x y c")
+    inference_result = inference_result.cpu().detach().numpy()
 
     # Define color mode according to number of channels in input images
     if temp_channels == 3:
@@ -83,7 +86,9 @@ if __name__ == "__main__":
         colormode = "L"
 
     # Save reconstructed images
-    for indx, uri in enumerate(datasets_uris):
-        im = Image.fromarray((np.squeeze(test_result[indx]) * 255).astype(np.uint8))
+    for indx in range(inference_result.shape[0]):
+        im = Image.fromarray(
+            (np.squeeze(inference_result[indx]) * 255).astype(np.uint8)
+        )
         im = im.convert(colormode)
         im.save(f"{args.output_dir}/reconstructed_{indx}.jpg")
